chore: #80 move tests into directories
standardized formatting
