+++ tests/mod.rs
@@ -1,3 +1,8 @@
 pub mod accounts;
+pub mod blockchain;
+pub mod blocks;
 pub mod common;
 pub mod node;
+pub mod txn;
+pub mod txn_pool;
+pub mod wallet;
