+++ tests/node/mod.rs
@@ -1,3 +1,4 @@
+// imports
 // local
 use posbc::node::{Node, Result};
 // test
